💾 Add: MyFactory 객체를 통해 My 뷰를 외부 메서드에서 생성
